+++ pelix/http/__init__.py
@@ -8,7 +8,7 @@ Defines the interfaces that must respect HTTP service implementations.
 :author: Thomas Calmant
 :copyright: Copyright 2026, Thomas Calmant
 :license: Apache License 2.0
-:version: 3.2.0
+:version: 3.2.1
 
 ..
 
@@ -38,7 +38,7 @@ from pelix.utilities import to_bytes
 # ------------------------------------------------------------------------------
 
 # Module version
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format