+++ pelix/misc/mqtt_client.py
@@ -10,7 +10,7 @@ Eclipse Foundation: see http://www.eclipse.org/paho
 :author: Thomas Calmant
 :copyright: Copyright 2026, Thomas Calmant
 :license: Apache License 2.0
-:version: 3.2.0
+:version: 3.2.1
 
 ..
 
@@ -43,7 +43,7 @@ from paho.mqtt.reasoncodes import ReasonCode
 # ------------------------------------------------------------------------------
 
 # Module version
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format