+++ pelix/misc/xmpp.py
@@ -9,7 +9,7 @@ This module depends on the slixmpp package: https://slixmpp.readthedocs.io/
 :author: Thomas Calmant
 :copyright: Copyright 2026, Thomas Calmant
 :license: Apache License 2.0
-:version: 3.2.0
+:version: 3.2.1
 
 ..
 
@@ -46,7 +46,7 @@ from pelix.utilities import EventData
 # ------------------------------------------------------------------------------
 
 # Module version
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format