+++ pelix/remote/discovery/zookeeper.py
@@ -9,7 +9,7 @@ PyPI), and a ZooKeeper server.
 :author: Thomas Calmant
 :copyright: Copyright 2026, Thomas Calmant
 :license: Apache License 2.0
-:version: 3.2.0
+:version: 3.2.1
 
 ..
 
@@ -48,7 +48,7 @@ from pelix.utilities import to_bytes, to_str
 # ------------------------------------------------------------------------------
 
 # Module version
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format