+++ pelix/remote/edef_io.py
@@ -9,7 +9,7 @@ specifications, section 122.8.
 :author: Thomas Calmant
 :copyright: Copyright 2026, Thomas Calmant
 :license: Apache License 2.0
-:version: 3.2.0
+:version: 3.2.1
 
 ..
 
@@ -39,7 +39,7 @@ from pelix.remote.beans import EndpointDescription
 # ------------------------------------------------------------------------------
 
 # Module version
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format