+++ pelix/rsa/providers/discovery/__init__.py
@@ -7,7 +7,7 @@ Discovery Provider API
 :author: Scott Lewis
 :copyright: Copyright 2020, Scott Lewis
 :license: Apache License 2.0
-:version: 3.2.0
+:version: 3.2.1
 
 ..
 
@@ -40,7 +40,7 @@ from pelix.rsa.endpointdescription import EndpointDescription
 # ------------------------------------------------------------------------------
 # Module version
 
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format