+++ pelix/rsa/providers/distribution/py4j.py
@@ -7,7 +7,7 @@ Py4j-based Distribution and Discovery Provider
 :author: Scott Lewis
 :copyright: Copyright 2020, Scott Lewis
 :license: Apache License 2.0
-:version: 3.2.0
+:version: 3.2.1
 
 ..
 
@@ -71,7 +71,7 @@ from pelix.rsa.providers.distribution import (
 # ------------------------------------------------------------------------------
 # Module version
 
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format