+++ pelix/services/__init__.py
@@ -6,7 +6,7 @@ Pelix OSGi-like services packages
 :author: Thomas Calmant
 :copyright: Copyright 2026, Thomas Calmant
 :license: Apache License 2.0
-:version: 3.2.0
+:version: 3.2.1
 
 ..
 
@@ -33,7 +33,7 @@ if TYPE_CHECKING:
     import pelix.ldapfilter as ldapfilter
 
 # Module version
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format