+++ pelix/shell/core.py
@@ -8,7 +8,7 @@ Provides the basic command parsing and execution support to make a Pelix shell.
 :author: Thomas Calmant
 :copyright: Copyright 2026, Thomas Calmant
 :license: Apache License 2.0
-:version: 3.2.0
+:version: 3.2.1
 
 ..
 
@@ -53,7 +53,7 @@ if TYPE_CHECKING:
 __all__ = ()
 
 # Module version
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format