+++ tests/framework/service_bundle.py
@@ -10,7 +10,7 @@ from pelix.constants import ActivatorProto, BundleActivator
 from pelix.framework import BundleContext
 from tests.interfaces import IEchoService
 
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 registered = False