+++ tests/ipopo/ipopo_fields_bundle.py
@@ -24,7 +24,7 @@ from pelix.ipopo.decorators import (
 
 # ------------------------------------------------------------------------------
 
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 SVC_A = "service.a"