+++ tests/rsa/endpoint_event_listener.py
@@ -11,7 +11,7 @@ from threading import RLock
 from pelix.ipopo.decorators import ComponentFactory, Provides
 from pelix.rsa.providers.discovery import EndpointEventListener
 
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)