+++ tests/shell/test_core.py
@@ -21,7 +21,7 @@ from pelix.shell import ShellCommandMethod, ShellCommandsProvider, ShellService,
 
 # ------------------------------------------------------------------------------
 
-__version_info__ = (3, 2, 0)
+__version_info__ = (3, 2, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format